requirements.txt
discord.py==2.1.0
openai==0.27.0
